Move blacklight_marc:marc generator to blacklight:marc:install
