+++ lib/generators/blacklight/marc/templates/config/SolrMarc/config-test.properties
