+++ lib/generators/blacklight/marc/templates/config/SolrMarc/config.properties
