+++ lib/generators/blacklight/marc/templates/config/SolrMarc/index.properties
