+++ lib/generators/blacklight/marc/templates/config/SolrMarc/index_scripts/dewey.bsh
