+++ lib/generators/blacklight/marc/templates/config/SolrMarc/translation_maps/composition_era_map.properties
