+++ lib/generators/blacklight/marc/templates/config/SolrMarc/translation_maps/country_map.properties
