+++ lib/generators/blacklight/marc/templates/config/SolrMarc/translation_maps/format_map.properties
