+++ lib/generators/blacklight/marc/templates/config/SolrMarc/translation_maps/instrument_map.properties
