+++ lib/generators/blacklight/marc/templates/config/SolrMarc/translation_maps/language_map.properties
